requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
python-dotenv>=1.0.1
email-validator>=2.1.0
uuid-utils>=0.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

